Remove deprecated constant
Delete constant that is not necessary anymore. The sources are configured in the input file.
